pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "coreason_etl_dgidb"
dynamic = ["version"]
description = "ETL pipeline for extracting and processing DGIdb drug-gene interaction data."
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "Akshaya M", email = "akshaya.movvar@coreason.ai" },
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dbt-core>=1.11.7",
    "dbt-postgres>=1.10.0",
    "dlt[postgres]>=1.23.0",
    "loguru>=0.7.2",
    "polars>=1.39.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
]
license = { file = "LICENSE" }
classifiers = [
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]

[dependency-groups]
dev = [
    "deepdiff",
    "deptry",
    "hypothesis",
    "mkdocstrings-python",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio",
    "pytest-benchmark",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.14.14",
    "syrupy",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20260107",
    "tzdata>=2025.3",
    "zensical",
]

[tool.hatch.build.targets.wheel]
packages = ["src/coreason_etl_dgidb"]

[tool.hatch.version]
source = "vcs"

[project.urls]
Homepage = "https://github.com/CoReason-AI/coreason_etl_dgidb"
Repository = "https://github.com/CoReason-AI/coreason_etl_dgidb"
Documentation = "https://github.com/CoReason-AI/coreason_etl_dgidb"

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "SIM", "RUF", "ARG", "C4", "PT", "TCH", "FA", "PIE", "RET", "PERF", "FURB", "LOG", "N", "A", "S"]
ignore = ["S101", "TC001", "TC002", "TC003", "UP037"]

[tool.mypy]
python_version = "3.14"
strict = true
disallow_untyped_defs = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
addopts = "-n auto --cov=src --cov-report=term-missing --cov-report=html --cov-fail-under=100"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["tests/*"]

[tool.deptry]
ignore = ["DEP002"]

[tool.uv]
override-dependencies = [
    "mashumaro[msgpack]>=3.20",
    "dbt-semantic-interfaces>=0.10.5"
]
